Move sub_sign to its sole use in multiplication
